% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternate.R
\name{alternate}
\alias{alternate}
\title{alternate}
\usage{
alternate(x, f, run_map, error_funs, error_args, return_raw)
}
\arguments{
\item{x}{A cfp_pfres or cfp_fgres model result.}
}
\description{
Alternate cfp_pfres / cfp_fgres models for sensitivit analysis and
more.
}
