% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{base_dat}
\alias{base_dat}
\title{Example cfp_dat object}
\format{
A \code{\link[=cfp_dat]{cfp_dat()}} object as a list with
\describe{
\item{profiles}{The profiles of the data.}
\item{gasdata}{The \link{gasdata} object.}
\item{soilphys}{The \link{soilphys} object.}
\item{layers_map}{The \link{layers_map} object.}
}
}
\usage{
base_dat
}
\description{
An example \code{\link[=cfp_dat]{cfp_dat()}} object that combines all other example data.
}
\keyword{datasets}
