% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_error.R
\name{bootstrap_error}
\alias{bootstrap_error}
\alias{bootstrap_error.cfp_pfres}
\alias{make_bootstrap_model}
\alias{make_bootstrap_model.cfp_pfres}
\alias{calculate_bootstrap_error}
\alias{calculate_bootstrap_error.cfp_pfres}
\title{Bootstrap-estimate of Flux-error}
\usage{
bootstrap_error(x, n_samples = 50, sd_x_ppm = NULL, n_replicates = NULL)

\method{bootstrap_error}{cfp_pfres}(x, n_samples = 50, sd_x_ppm = NULL, n_replicates = NULL)

make_bootstrap_model(x, n_samples = 50, sd_x_ppm = NULL, n_replicates = NULL)

\method{make_bootstrap_model}{cfp_pfres}(x, n_samples = 50, sd_x_ppm = NULL, n_replicates = NULL)

calculate_bootstrap_error(x, y)

\method{calculate_bootstrap_error}{cfp_pfres}(x, y)
}
\arguments{
\item{x}{A \code{cfp_pfres} model result from a call to \code{\link[=pro_flux]{pro_flux()}}.}

\item{n_samples}{The number of samples to take in the bootstrapping.}

\item{sd_x_ppm}{An optional estimate of the standard deviation of x_ppm.}

\item{n_replicates}{The number of replicates to be generated if sd_x_ppm is
set.}

\item{y}{The result of the bootstrap model.}
}
\value{
x with added columns DELTA_flux and DELTA_prod as an estimate
of the error of of the corresponding columns in the same units.
}
\description{
Bootstrap-estimate of Flux-error
}
\details{
\code{\link[=bootstrap_error]{bootstrap_error()}} is mostly a wrapper around two functions that can also be
run separately.

In \code{\link[=make_bootstrap_model]{make_bootstrap_model()}}, the \code{gasdata} concentration data is
resampled for every depth and profile a total number of \code{n_samples}.
This is done by randomly sampling the observations at each depth without
changing the number of observations but while allowing replacing. Each newly
sampled profile is identifiable by the added \code{bootstrap_id} column which
is also added to \code{id_cols}.

After this new model is run again, the bootrstap error is caculated in
\code{\link[=calculate_bootstrap_error]{calculate_bootstrap_error()}}. This is the standard deviation of the
production and flux parameters across all bootstrapped model runs and is
calculated for each profile and layer of the original model.
These are added to the \code{PROFLUX} data of the original model and can
thereby be extracted by \code{\link[=efflux]{efflux()}}.
}
\examples{
PROFLUX <- pro_flux(ConFluxPro::base_dat)
PROFLUX_BSE <- bootstrap_error(PROFLUX)
efflux(PROFLUX_BSE)
}
