% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flux.R
\name{calculate_flux}
\alias{calculate_flux}
\title{calculate_flux}
\usage{
calculate_flux(gasdata, soilphys, layers_map, gases, modes, param, funs)
}
\arguments{
\item{gasdata}{(dataframe) The gasdata dataframe}

\item{soilphys}{(dataframe) The soilphys dataframe after discretisation and completion.}

\item{layers_map}{(dataframe) containing the following parameters: "layer"=name of the layer;
"upper"=upper limit of layer in cm; "lower" = lower limit of the layer in cm;}

\item{modes}{(character) A character vector specifying mode(s) for dcdz calculation. Can be "LL","LS","EF".}

\item{param}{(character) A vector containing the the parameters of soilphys, for which means should be calculated}

\item{funs}{(character) A vector defining the type of mean to be used. One of "arith" or "harm"}
}
\value{
FLUX
}
\description{
This function takes the soilphys-Dataframe as well as the gasdata-dataframe and calculates fluxes accordingly.
}
\examples{

}
