% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_fgmod.R
\name{cfp_fgmod}
\alias{cfp_fgmod}
\title{cfp_fgmod}
\usage{
cfp_fgmod(x, gases, modes, param, funs)
}
\arguments{
\item{x}{A \code{cfp_dat} object with all the necessary input datasets.}

\item{gases}{(character) A character vector defining the gases for which
fluxes shall be calulated.}

\item{modes}{(character) A character vector specifying mode(s) for dcdz
calculation. Can be "LL","LS","EF".}

\item{param}{(character) A vector containing the the parameters of soilphys,
for which means should be calculated, must contain "c_air" and "DS", more
parameters help interpretation}

\item{funs}{(character) A vector defining the type of mean to be used. One of
"arith" or "harm"}
}
\description{
An S3 class for \code{fg_flux()} models. The class inherits from
cfp_dat and adds any model specific parameters.
}
