% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_gasdata.R
\name{cfp_gasdata}
\alias{cfp_gasdata}
\alias{cfp_gasdata.data.frame}
\alias{cfp_gasdata.cfp_dat}
\title{cfp_gasdata}
\usage{
cfp_gasdata(x, ...)

\method{cfp_gasdata}{data.frame}(x, id_cols, ...)

\method{cfp_gasdata}{cfp_dat}(x, ...)
}
\arguments{
\item{x}{A \code{data.frame} with the following columns:
\describe{
\item{gas}{The gas of that observation.}
\item{depth (cm)}{The depth of the observation.}
\item{x_ppm (ppm)}{The observed concentration of that gas.}
\item{any of \code{id_cols}}{All id_cols that identify one profile uniquely.}
}}
}
\value{
cfp_gasdata
}
\description{
A function to create and validate cfp_gasdata objects. A gasdata
object is characterised by the following criteria: Each row is one observation
of the concentration of a gas at a depth and can be attributed to a distinct
profileidentified by any \code{id_cols} columns.
}
