% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_parameter.R
\name{cfp_parameter}
\alias{cfp_parameter}
\title{Get parameter descriptions and units}
\usage{
cfp_parameter(x = NULL)
}
\arguments{
\item{x}{Any object or data.frame to match the parameters to, or a character
vector of parameter names.}
}
\description{
Function to access parameter descriptions and units
used in \code{ConFluxPro}
}
