% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_dat.R, R/cfp_profile.R
\name{split_by_prof.cfp_profile}
\alias{split_by_prof.cfp_profile}
\alias{cfp_profile}
\title{cfp_profile}
\usage{
\method{split_by_prof}{cfp_profile}(x)

cfp_profile(x, id_cols = NULL)
}
\arguments{
\item{x}{A \code{data.frame}}

\item{id_cols}{Column names in data.frame that uniquely identify each profile.}
}
\description{
A central S3 class that defines a \code{data.frame} where columns given in
\code{id_cols} define distinct soil profiles.
}
