% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_run_map.R
\name{cfp_run_map}
\alias{cfp_run_map}
\title{cfp_run_map}
\usage{
cfp_run_map(x, ...)
}
\arguments{
\item{x}{An object from which to extract a run_map.}

\item{...}{Additional arguments to form a valid cfp_run_map object.}
}
\description{
An S3 class for run maps generated by a call to \link{run_map}() to
be used in alternate(). Either create a new run map from a cfp_pfres or
cfp_fgres model or extract an existing run_map from an cfp_altres object.
}
