% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_soilphys.R
\name{cfp_soilphys}
\alias{cfp_soilphys}
\alias{cfp_soilphys.cfp_dat}
\alias{cfp_soilphys.data.frame}
\title{cfp_soilphys}
\usage{
cfp_soilphys(x, ...)

\method{cfp_soilphys}{cfp_dat}(x, ...)

\method{cfp_soilphys}{data.frame}(x, id_cols, ...)
}
\arguments{
\item{x}{(data.frame) That defines the layers for which the
production or flux is modeled. Note that some parameters can also be provided
directly to the function call instead (see Details).
\itemize{
\item \code{id_cols} the relevant id_cols (see below)
\item \code{gas}, the gas that is modelled.
\item \code{upper}, \code{lower} the upper and lower boundaries
of each layer
\item \code{lowlim}, \code{highlim} as the lower and upper limits of the
production rate to be modeled in \eqn{\mu~mol~m^{-3}}
\item the parameter \code{layer_couple}, that indicates how
strongly the layer should be linked to the one below it (0 for no coupling)
}}

\item{soilphys}{A data.frame with (at least) the following columns:
\describe{
\item{upper (cm) }{The upper bound of each step.}
\item{lower (cm)}{The lower bound of each step.}
\item{gas}{The gas of that step.}
\item{DS (\eqn{m^2 s^-1})}{The specific diffusion coefficient of that gas in that step.}
\item{c_air (\eqn{mol m^-3})}{The number density of air in that step.}
\item{any of \code{id_cols}}{All id_cols that identify one profile uniquely.}

}}
}
\value{
cfp_soilphys
}
\description{
Create and validate cfp_soilphys objects. A cfp_soilphys object has the following
characteristics. Each profile is uniquely identifiable by its \code{id_cols} without duplicates.
The data is upper/lower consistent, meaning that the steps of each profile cover a range
from min(lower) to max(upper) without gaps, overlap or duplicates.
}
