% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repetitions_layered_profile.R
\name{check_matching_repetitions}
\alias{check_matching_repetitions}
\title{Repetitions in layered profiles}
\usage{
check_matching_repetitions(x, rep_cols)
}
\arguments{
\item{x}{A \code{cfp_layered_profile} object.}

\item{rep_cols}{A character vector of columns that define repetitions of
profiles within \code{x}. Must be a subset of the \code{id_cols} of x.}
}
\description{
\code{check_matching_repetitions()} compares the structure of layered profile objects. A \code{\link[=cfp_layered_profile]{cfp_layered_profile()}}
object is split along parameter(s) defined in \code{rep_cols}. Returns TRUE
if all profiles match in their structure.
}
\keyword{internal}
