% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunk_lapply}
\alias{chunk_lapply}
\title{utility functions}
\usage{
chunk_lapply(X, FUN, fun_process = function(a) {
    a
}, n_per_chunk = 1)
}
\description{
A group of functions that help in the internals of the package.
}
