% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_fgmod.R, R/cfp_pfmod.R, R/coercion.R
\name{as_cfp_fgmod}
\alias{as_cfp_fgmod}
\alias{as_cfp_pfmod}
\alias{coercion}
\title{coercion}
\usage{
as_cfp_fgmod(x)

as_cfp_pfmod(x)
}
\arguments{
\item{x}{An object which should be coerced}
}
\value{
Either a \link{cfp_fgmod} or \link{cfp_pfmod} model frame
}
\description{
These functions help coerce different object types in
ConFluxPro.
}
\section{Functions}{
\itemize{
\item \code{as_cfp_fgmod()}: to cfp_fgmod

\item \code{as_cfp_pfmod()}: to cfp_pfmod

}}
\examples{
PROFLUX <- pro_flux(base_dat)
as_cfp_pfmod(PROFLUX)

FLUX <- fg_flux(base_dat)
as_cfp_fgmod(FLUX)


}
\keyword{internal}
