% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\alias{combine_models.cfp_altres}
\alias{combine_models.list}
\title{Combine Models}
\usage{
combine_models(x)

\method{combine_models}{cfp_altres}(x)

\method{combine_models}{list}(x)
}
\description{
A short description...
}
