% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_soilphys.R
\name{complete_soilphys}
\alias{complete_soilphys}
\title{complete_soilphys}
\usage{
complete_soilphys(soilphys, DSD0_formula = "a*AFPS^b", gases, overwrite = F)
}
\arguments{
\item{soilphys}{(dataframe) the soilphys dataframe}

\item{DSD0_formula}{(character) A character vector defining the way DSD0 should be calculated. Must refer to existing variables in soilphys. See examples below}

\item{gases}{(character) A character vector defining the gases for which to calculate D0 and DS.}

\item{overwrite}{(logical) If true, existing columns are overwritten.}
}
\description{
This function completes the soilphys dataset by calculating different parameters if necessary, as long as
all required parameters are available.
}
\examples{
soilphys <- complete_soilphys(soilphys,DSD0 = "a*AFPS^b")

}
