% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConFluxPro-package.R
\docType{package}
\name{ConFluxPro-package}
\alias{ConFluxPro}
\alias{ConFluxPro-package}
\title{ConFluxPro: A Toolkit for Soil Gas Analysis}
\description{
A suite to model soil gas fluxes with the Flux-Gradient Method. It includes functions for data handling, a forward and an inverse model for flux estimation and methods for calibration and uncertainty estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://confluxpro.valentingartiser.de/}
  \item \url{https://github.com/valentingar/ConFluxPro}
  \item \url{https://valentingar.github.io/ConFluxPro/}
  \item Report bugs at \url{https://github.com/valentingar/ConFluxPro/issues}
}

}
\author{
\strong{Maintainer}: Valentin Gartiser \email{code@valentingartiser.de} (\href{https://orcid.org/0000-0001-5320-374X}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Martin Maier \email{martin.freudiger@uni-goettingen.de} (\href{https://orcid.org/0000-0002-7959-0108}{ORCID}) [contributor]
  \item Verena Lang \email{verena.lang@forst.bwl.de} [contributor]
}

}
\keyword{internal}
