% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D0_massman.R
\name{D0_massman}
\alias{D0_massman}
\title{D0_Massman}
\usage{
D0_massman(gas, Temp, p)
}
\arguments{
\item{gas}{(character) One of "CO2","CH4","N2O","O2","N2"}

\item{p}{(numeric) pressure in hpa}

\item{t}{(numeric) temperature in °C}
}
\value{
D0 in m^2/s
}
\description{
This function calculates the Diffusion coefficients of different gases in air at set
temperature and pressure.
}
\references{
Massman, W. J. A review of the molecular diffusivities of H2O,
CO2, CH4, CO, O3, SO2, NH3, N2O, NO, and NO2 in air, O2 and N2 near STP.
Atmospheric Environment 1998, 32(6), 1111–1127
}
