% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcdz_ef.R
\name{dcdz_ef}
\alias{dcdz_ef}
\title{dcdz_ef}
\usage{
dcdz_ef(df, topheight, mode)
}
\arguments{
\item{df}{(dataframe) the gasdata dataframe, filtered to one profile (e.g. 1 day & one Plot ).}

\item{topheight}{(numeric) Depth at the Atmosphere/Soil interface. (e.g. Height of Humus)}

\item{mode}{(character) One of ("EF","LR").}
}
\value{
df (dataframe) Following columns:

mode (character) the used gradient method.

topheight (numeric) the topheight of that profile

dcdz_ppm (numeric) concentration gradient in ppm/cm.

dcdz_sd (numeric) standard error in ppm/cm.

r2 (linearised) R^2 for the model.
}
\description{
This function calculates concentration gradients for complete profiles with different approaches
}
\examples{

}
