% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efflux_extrap.R
\name{efflux_extrap}
\alias{efflux_extrap}
\title{efflux_extrap}
\usage{
efflux_extrap(FLUX, gases = "all", method, layers = NA, modename = NA, id_cols)
}
\arguments{
\item{FLUX}{(dataframe) the FLUX dataframe}

\item{gases}{(character vector) A (vector) of gases to be extrapolated as
stated in the function call. Fefaults to all gases.}

\item{method}{(character) A string defining the method to be used for
 extrapolation. Must be called multiple times for different methods. \cr

 Possible methods are:
 \describe{
 \item{lm}{implements a linear model
 approach. Here a linear model is fit to flux ~ depth and the value for the
 surface estimated}
 \item{linextrap}{implements a linear extrapolation
 approach. Here, two layer names must be given using the control parameters
 below, that will then be used for an linear extrapolation. }
 }}

\item{layers}{(character vector) layers to be used in the linextrap-approach.}

\item{modename}{(character) A character defining the value of the variable

 "mode" in the returned dataframe.}

\item{id_cols}{(character) A character vector defining the columns that
uniquely identify one profile.}
}
\value{
EFFLUX
}
\description{
This function extrapolates fluxes calculated with the layers
 approach to the surface with different methods. To see the different
 approaches implemented check the "method" parameter below. The values are
 grouped by Plot,Date,gas,mode.
}
\examples{
efflux_extrap(FLUX,
                        gases = c("CO2","O2"),
                        method = "linextrap",
                        layers = c("MIN1","MIN2"),
                        modename = "Hirano",
                        id_cols = c("Date"))

}
\seealso{
Other FLUX: 
\code{\link{calculate_flux}()},
\code{\link{dcdz_ef}()},
\code{\link{dcdz_layered}()}
}
\concept{FLUX}
