% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{evaluate_models}
\alias{evaluate_models}
\alias{evaluate_models.cfp_altres}
\title{evaluate_models}
\usage{
evaluate_models(x, eval_funs = NULL, eval_args = NULL)

\method{evaluate_models}{cfp_altres}(x, eval_funs = NULL, eval_args = NULL)
}
\description{
Evaluate the model runs produced by a call to \code{\link[=alternate]{alternate()}} with user-defined
error functions.
}
