% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R, R/cfp_altres.R, R/cfp_dat.R,
%   R/cfp_fgmod.R, R/cfp_pfmod.R, R/cfp_run_map.R
\name{extractors}
\alias{extractors}
\alias{cfp_og_model}
\alias{cfp_og_model.cfp_altres}
\alias{cfp_id_cols}
\alias{cfp_gases}
\alias{cfp_modes}
\alias{cfp_param}
\alias{cfp_funs}
\alias{cfp_zero_flux}
\alias{cfp_zero_limits}
\alias{cfp_DSD0_optim}
\alias{cfp_evenness_factor}
\alias{cfp_known_flux_factor}
\alias{cfp_runmap_type}
\alias{cfp_params_df}
\alias{cfp_n_runs}
\alias{cfp_layers_different}
\alias{cfp_layers_from}
\alias{cfp_layers_altmap}
\title{extractors}
\usage{
cfp_og_model(x)

\method{cfp_og_model}{cfp_altres}(x)

cfp_id_cols(x)

cfp_gases(x)

cfp_modes(x)

cfp_param(x)

cfp_funs(x)

cfp_zero_flux(x)

cfp_zero_limits(x)

cfp_DSD0_optim(x)

cfp_evenness_factor(x)

cfp_known_flux_factor(x)

cfp_runmap_type(x)

cfp_params_df(x)

cfp_n_runs(x)

cfp_layers_different(x)

cfp_layers_from(x)

cfp_layers_altmap(x)
}
\arguments{
\item{x}{An object from which to extract the information.}
}
\description{
These functions extract components from
different objects that can be created in ConFluxPro.
}
