% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_flux.R
\name{fg_flux}
\alias{fg_flux}
\title{fg_flux}
\usage{
fg_flux(x, ...)
}
\arguments{
\item{x}{A \code{cfp_dat} object with all the necessary input datasets.}

\item{...}{
  Arguments passed on to \code{\link[=cfp_fgmod]{cfp_fgmod}}
  \describe{
    \item{\code{gases}}{(character) A character vector defining the gases for which
fluxes shall be calulated.}
    \item{\code{modes}}{(character) A character vector specifying mode(s) for dcdz
calculation. Can be "LL","LS","EF".}
    \item{\code{param}}{(character) A vector containing the the parameters of soilphys,
for which means should be calculated, must contain "c_air" and "DS", more
parameters help interpretation}
    \item{\code{funs}}{(character) A vector defining the type of mean to be used. One of
"arith" or "harm"}
  }}
}
\description{
This function takes a valid input dataset in cfp_dat
and calculates fluxes accordingly. Fluxes are calculated for each layer defined
in layers map and are given in mumol/m^2/s.
}
