% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasdata}
\alias{gasdata}
\title{Soil CO2 concentrations}
\format{
A tibble with 312 rows and 6 variables:
\describe{
\item{site}{name of the site}
\item{Date}{Date in the format "YYYY-MM-DD"}
\item{depth}{depth from mineral soil in cm}
\item{repetition}{id for which repetition in each depth}
\item{x_ppm}{concentration, in ppm}
\item{gas}{name of the gas}
}
}
\usage{
gasdata
}
\description{
A synthetic dataset of soil CO2 concentrations at
two sites over a one-year period.
}
\keyword{datasets}
