% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputclasses.R
\name{gdat}
\alias{gdat}
\title{input_classes}
\usage{
gdat(tbl, id_cols)
}
\arguments{
\item{x}{sosso}
}
\description{
This provides the framework for multiple input data classes that
make it easier to set up a ConFluxPro model by checking for data integrity first.
}
