% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layers_map}
\alias{layers_map}
\title{layers_map}
\format{
A data.frame with 4 rows and 3 variables:
\describe{
\item{site}{name of the site}
\item{upper}{upper limit for layer in cm}
\item{lower}{lower limit for layer in cm}
}
}
\usage{
layers_map
}
\description{
An example dataset for \code{layers_map} that devides each \code{site} into
two layers.
}
\keyword{datasets}
