% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_profiles.R
\name{n_profiles}
\alias{n_profiles}
\title{n_profiles}
\usage{
n_profiles(x)
}
\arguments{
\item{x}{A \code{cfp_dat} object with all the necessary input datasets.}
}
\description{
Get the number of profiles in cfp objects.
}
