% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFres.R
\name{new_PFres}
\alias{new_PFres}
\title{PFres}
\usage{
new_PFres(
  tbl = list(),
  layers_map = list(),
  id_cols = character(),
  storage_term = logical(),
  zero_flux = logical(),
  zero_limits = numeric(),
  known_flux_factor = numeric(),
  Ds_optim = logical(),
  evenness_factor = numeric()
)
}
\description{
This is the central S3 class generated by pro_flux() that
stores all the additional function arguments used in the pro_flux()
function call, except for \code{gasdata} \code{soilphys} and \code{known_flux}.

It behaves just like a data.frame but stores additional information
in the attributes. This enables the extraction of additional information
such as efflux and also helps keeping track of different models, similar
to the \code{lm} class.
}
