% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_override.R
\name{offset_override}
\alias{offset_override}
\title{offset_override}
\usage{
offset_override(corr_map, SAMPLE_NO, section)
}
\arguments{
\item{corr_map}{(dataframe) The corr_map dataframe created by
offset_subsetting}

\item{SAMPLE_NO}{(character) A vector of the SAMPLE_NO to change section}

\item{section}{(vector, numeric) A vector in the same order as SAMPLE_No
specifying the target section.}
}
\value{
corr_map (dataframe) The corr_map dataframe used in the
  offset_correction function with overridden SAMPLE_NO.
}
\description{
This function helps overriding the Date-based subsetting of
  offset_subsetting(). It modifies an existing corr_map by editing the
  section of a list of SAMPLE_NO. Check the section numbering of corr_map
  first. If a new section is manually introduced this way, a mode will have
  to be added manually.
}
\examples{
{
data("gasdata")

library(dplyr)

gasdata <- gasdata \%>\%
  mutate(depth_cat = ifelse(depth>0,"HU","MIN"),
         SAMPLE_NO = row_number())

cmap <-
  offset_subsetting(gasdata,
                    gas = "CO2",
                    depth_cal = "HU",
                    start = "2021-01-01",
                    end = "2022-01-01",
                    mode = "const")
new_section <-
  cmap$SAMPLE_NO[cmap$Date == "2021-01-01"]

cmap <-
  offset_override(cmap,new_section,2)

cmap$mode[cmap$section==2] <- "const"

offset_correction(gasdata,
                  corr_map = cmap,
                  gases = "CO2",
                  gases_std = 400,
                  depth_cal = "HU")

}

}
\seealso{
offset_subsetting

offset_correction

Other gasdata: 
\code{\link{balance_correction}()},
\code{\link{offset_correction}()},
\code{\link{offset_subsetting}()},
\code{\link{series_cleaner}()}
}
\concept{gasdata}
