% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_extract.R
\name{pf_efflux}
\alias{pf_efflux}
\title{proflux_extract}
\usage{
pf_efflux(PROFLUX, ...)
}
\arguments{
\item{PROFLUX}{The data.frame returned by pro_flux()}

\item{target}{Character string defining the type of parameter to be
extracted / aggregated.
One of:
\describe{
  \item{efflux}{Extract the efflux at the atmosphere/soil interface.
This is the flux at the topmost position.}
\item{prod_layer}{Extract the total production per layer defined in the function call
of pro_flux()}
}}
}
\value{
data.frame
}
\description{
Utility function for the fast extraction and aggregation of
parameters exported by pro_flux
}
