% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_parallel_automatic.R
\name{plan_parallel_automatic}
\alias{plan_parallel_automatic}
\title{Automatically plan for parallel computing in cfp.}
\usage{
plan_parallel_automatic(x, number_of_workers = future::availableCores())
}
\arguments{
\item{x}{A cfp_dat object.}

\item{number_of_workers}{Number of session to create in total.
Defaults to number of available cores.}
}
\description{
Automatically plan for parallel computing in cfp.
}
