% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profile.R
\name{plot_profile}
\alias{plot_profile}
\title{plot_profile}
\usage{
plot_profile(x)
}
\arguments{
\item{x}{}
}
\description{
Plot ConFluxPro profiles using ggplot. Supported objects:
\describe{
\item{cfp_soilphys}{Displays TPS, SWC and AFPS, as well as values of Ds and Temperature.}
\item{cfp_pfres}{Displays TPS, SWC and AFPS, as well as production and measured and modelled gas concentrations.}
}
}
