% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_mod.R
\name{prod_mod}
\alias{prod_mod}
\title{prod_mod}
\usage{
prod_mod(df, F0, C0)
}
\arguments{
\item{df}{(dataframe) The input dataframe, sorted lowest depth to highest,
containg the following columns: \n
height: the height of each step in m
prod: the production assinged to this step
DS: the diffusion coefficient DS}

\item{F0}{(numeric) Incoming flux to lowest step in mumol/s/m^2}

\item{C0}{(numeric) Concentration at the lower end of the lowest step.}
}
\description{
Function calculates flux and concentration of a gas based on
fixed production and known diffsuion coefficient DS. The basis is a stepwise
calculation beginning from the bottom-most layer with incoming flux F0 and
known Concentration C0 at the lower end of the step.
}
\examples{
prod_mod(df,F0=0,C0=1200)

}
