% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_alternate.R
\name{proflux_alternate}
\alias{proflux_alternate}
\title{proflux_alternate}
\usage{
proflux_alternate(
  PROFLUX,
  params,
  facs,
  params_map,
  no_confirm = F,
  DSD0_formula
)
}
\arguments{
\item{PROFLUX}{An object of class PFres, the result
of a call to \code{\link{pro_flux}}}

\item{params}{A vector of characters which parameters to alternate.
Can be the character of any column in \code{soilphys} or "topheight".
If "topheight" is called, the height of the soil/atmosphere interface will
be altered. This means, that the corresponding values in \code{gasdata},
\code{soilphys} and \code{layers_map} will be adjusted.

The parameter(s) provided will be varied along the \code{facs} provided.
Then, \code{\link{complete_soilphys}} is run. This means, that it does not make
sense to alter, e.g. "AFPS" as it is re-calculated in that function call.}

\item{facs}{A vector of factors each of params will be multiplied by.
This may be something like \code{seq(0.9,1.1,0.05)} if values should be varied between
90 \% and 110 \% of their original value.}

\item{params_map}{A \code{data.frame} containing any of \code{id_cols} of
PROFLUX, as well as \code{upper} and \code{lower}. For each layer enclosed by
"upper" and "lower", the parameters in "params" will be varied independently.
This means, that the more layer per combination of id_cols are provided, the
more variations will be tested out.}

\item{no_confirm}{If the function is used in a script, set this to TRUE to confirm
the number of profiles without separate user input. Please ensure separately if
the function will evaluate in a reasonable amount of time.}
}
\description{
Functions to extract datasets from the result of
a \code{\link{pro_flux}} model, a PFres object, to modify
parameters and rerun the model.
}
