% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_error.R
\name{proflux_error}
\alias{proflux_error}
\title{proflux_error}
\arguments{
\item{x}{A cfp_pfres object, that is returned by a call
to pro_flux()}

\item{param_cols}{The columns that, together, define different parameters (e.g. different gases)
for which NRMSEs should be calculated separately (e.g. "gas").}
}
\description{
A set of functions that can be called on an
cfp_pfres object (the result of a call to pro_flux) to assess
the quality of the model.
}
