% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_summarise.R
\name{proflux_summarise}
\alias{proflux_summarise}
\alias{proflux_summarize}
\alias{proflux_summarise.data.frame}
\alias{proflux_summarise.PFres}
\title{proflux_summarise}
\usage{
proflux_summarise(PROFLUX, ...)

proflux_summarize(PROFLUX, ...)

\method{proflux_summarise}{data.frame}(PROFLUX, layers_map, id_cols, ...)

\method{proflux_summarise}{PFres}(PROFLUX, ...)
}
\arguments{
\item{PROFLUX}{Either an object of class \code{\link{PFres}}
(see \code{\link{pro_flux}}), or a data.frame following the same convention.}

\item{layers_map}{(dataframe) That defines the layers of homogeneous
production as well as the upper and lower limits of production rate. \ Must
include
\itemize{
\item the relevant id_cols (see below)
\item the upper
and lower boundaries of each layer (upper, lower)
\item upper and lower
limits of the production rate to be modeled in \eqn{\mu mol m^{-3}}
(highlim, lowlim)
\item the parameter layer_couple, that indicates how
strongly the layer should be linked to the one below it (0 for no coupling)
}}

\item{id_cols}{(character vector) The names of the columns that together
uniquely identify one profile. Must be present in gasdata.}
}
\description{
Summarises key parameters of a PFres object
(produced by a call to \code{\link{pro_flux}}) based on the id_cols of its layers_map.
}
