% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrmse.R
\name{rmse}
\alias{rmse}
\alias{nrmse}
\title{(N)RMSE}
\usage{
rmse(a, b)

nrmse(a, b, normer)
}
\arguments{
\item{a, b}{numeric vectors of same length to be compared}

\item{normer}{a character string defining the type of
normalization to be applied. Can be one of
\describe{
\item{mean}{the arithmetic mean of a}
\item{sd}{the standard deviation of a}
\item{range}{the difference between the range of a}
\item{IQR}{the difference between the interquantile range of a}
}}
}
\description{
(N)RMSE

Calculate the (normalized) root-mean-square-error of
two vectors.
}
