% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{scale_min_median}
\alias{scale_min_median}
\title{scale_min_median}
\usage{
scale_min_median(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
Scale a vector between its minimum and median.
}
