% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{Get season of a date-time}
\usage{
season(d)
}
\arguments{
\item{d}{(Date) Any date object}
}
\value{
A character vector the same length as d
}
\description{
A simple function to return a character (-vector)
of the season from a Date (-vector).
Months:
\describe{
\item{spring}{3-5}
\item{summer}{6-8}
\item{fall}{9-11}
\item{winter}{12-2}
}
}
\examples{
season(as.Date(c("1955-01-15","1985-06-15","2015-10-15")))

}
