% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soildiff}
\alias{soildiff}
\title{Example soil diffusion models}
\format{
A tibble with 8 rows and 6 variables:
\describe{
\item{site}{name of the site}
\item{upper}{upper limit for layer in cm}
\item{lower}{lower limit for layer in cm}
\item{TPS}{total pore space as fraction of volume}
\item{a}{diffusion-model fit parameter a}
\item{b}{diffusion-model fit parameter b}
}
}
\usage{
soildiff
}
\description{
A synthetic dataset of soil total pore space and
diffusion models after the general formula a*AFPS^b.
}
\keyword{datasets}
