% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilphys}
\alias{soilphys}
\title{soilphys}
\format{
A tibble with 120 rows and 4 variables:
\describe{
  \item{site}{name of the site}
  \item{Date}{Date in the format "YYYY-MM-DD"}
  \item{depth}{depth in cm}
  \item{t}{temperature in °C}
}
}
\usage{
soilphys
}
\description{
An example dataset for \code{soilphys} based on the sets
\code{soiltemp}, \code{soilwater} and \code{soildiff}
}
\keyword{datasets}
