% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soiltemp}
\alias{soiltemp}
\title{Soil temperature}
\format{
A tibble with 120 rows and 4 variables:
\describe{
\item{site}{name of the site}
\item{Date}{Date in the format "YYYY-MM-DD"}
\item{depth}{depth in cm}
\item{t}{temperature in °C}
}
}
\usage{
soiltemp
}
\description{
A synthetic dataset of soil temperature at
discrete depths. The dates correspond to \code{gasdata}.
}
\keyword{datasets}
