% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{soilwater}
\alias{soilwater}
\title{Example soil water content}
\format{
A tibble with 180 rows and 5 variables:
\describe{
\item{site}{name of the site}
\item{Date}{Date in the format "YYYY-MM-DD"}
\item{upper}{upper limit for layer in cm}
\item{lower}{lower limit for layer in cm}
\item{SWC}{soil water content as fraction of volume}
}
}
\usage{
soilwater
}
\description{
A synthetic dataset of soil water content in a
layered structure. The dates correspond to \code{gasdata}.
}
\keyword{datasets}
