% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_bar.R
\docType{data}
\name{Statvariable_bar}
\alias{Statvariable_bar}
\title{variable bar}
\format{
An object of class \code{variable_bar} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
Statvariable_bar
}
\description{
A custom ggplot2 stat that implements a variable-width bar plot.
}
\keyword{datasets}
