% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unique_gases.R
\name{unique_gases}
\alias{unique_gases}
\title{Get the unique gases of an object}
\usage{
unique_gases(x)
}
\arguments{
\item{x}{the object to extract the gases from.}
}
\value{
A character vector of gases in that object.
}
\description{
Get the gases from a CFP object.
}
\examples{
unique_gases(base_dat)

data.frame(gas = c("CO2", "CH4")) |>
cfp_profile(id_cols = "gas") |>
unique_gases()

}
