% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputclasses.R
\name{cfp_gasdata}
\alias{cfp_gasdata}
\title{input_classes}
\usage{
cfp_gasdata(gasdata, id_cols)
}
\arguments{
\item{x}{sosso}
}
\description{
This provides the framework for multiple input data classes that
make it easier to set up a ConFluxPro model by checking for data integrity first.
}
