% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_layered_profile.R
\name{cfp_layered_profile}
\alias{cfp_layered_profile}
\title{cfp_layered_profile}
\usage{
cfp_layered_profile(x, id_cols = NULL)
}
\arguments{
\item{x}{A \code{data.frame} with columns \code{upper} and \code{lower}.}

\item{id_cols}{Column names in data.frame that uniquely identify each profile.}
}
\description{
A subclass of \code{\link{cfp_profile}()} where each profile consists of
layers that are defined by their \code{upper} and \code{lower}
boundary without gaps or duplicates.
}
\details{
\code{upper} and \code{lower} define the upper and lower bounds of
each layer in cm. Higher values lay on top of lower values.
}
