% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_layers_map.R
\name{cfp_layers_map}
\alias{cfp_layers_map}
\title{cfp_layers_map}
\usage{
cfp_layers_map(
  layers_map,
  id_cols,
  gas = NULL,
  lowlim = NULL,
  highlim = NULL,
  layer_couple = NULL
)
}
\arguments{
\item{layers_map}{(dataframe) That defines the layers of homogeneous
production as well as the upper and lower limits of production rate. \ Must
include
\itemize{
\item the relevant id_cols (see below)
\item the upper
and lower boundaries of each layer (upper, lower)
\item upper and lower
limits of the production rate to be modeled in \eqn{\mu mol m^{-3}}
(highlim, lowlim)
\item the parameter layer_couple, that indicates how
strongly the layer should be linked to the one below it (0 for no coupling)
}}

\item{id_cols}{(character vector) The names of the columns that together
uniquely identify one profile.}

\item{gas}{(character vector) A vector of gas names to be added to layers_map.
The input layers_map is then repeated for each gas.}

\item{lowlim}{(numeric vector) A vector the same length as gas with the lower limit
of possible production allowed in \code{pro_flux} models.}

\item{highlim}{(numeric vecotr)  A vector the same length as gas with the upper limit
of possible production allowed in \code{pro_flux} models.}

\item{layer_couple}{(numeric_vector) A vector the same length as gas that indicates how
strongly the layer should be linked to the one below it (0 for no coupling)}
}
\description{
A function to conveniently create a vaild cfp_layers_map object
with options to add more gases and corresponding parameters for \code{\link{pro_flux}} models.
Please note that only \code{layers_map} and code{id_cols} are obligatory. Adding the necessary
columns to \code{layers_map} beforehand allows for more fine-tuned control of models.
}
