% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_pfres.R
\name{cfp_pfres}
\alias{cfp_pfres}
\title{cfp_pfres}
\usage{
cfp_pfres(x, y)
}
\arguments{
\item{x}{A valid \code{cfp_pfmod} object}

\item{y}{The corresponding PROFULX \code{data.frame}.}
}
\value{
A \code{cfp_pfres} object. This inherits from \code{\link[=cfp_pfmod]{cfp_pfmod()}}.
}
\description{
A function to create an object of class cfp_pfres. This is the
central result class generated by running \code{pro_flux()}. Intended for
internal use only.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
PROFLUX <- pro_flux(ConFluxPro::base_dat)
cfp_pfres(
  cfp_pfmod(ConFluxPro::base_dat),
  PROFLUX$PROFLUX
)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
