% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_run_map.R
\name{cfp_run_map}
\alias{cfp_run_map}
\title{cfp_run_map}
\usage{
cfp_run_map(
  x,
  id_cols,
  params,
  method,
  type,
  n_runs,
  layers_different,
  runmap_type,
  params_df
)
}
\arguments{
\item{params}{A named list of numeric vectors. Names indicate column names
in soilphys, vectors either distinct values (method permutation) or limits (method random).}

\item{type}{A vector of length param indicating what the values in params
represent. One of
\describe{
\item{abs}{Absolute values that are applied as-is.}
\item{factor}{Factors to be multiplied with the original values.}
\item{addition}{Factors to be added to the original values.}
}}

\item{n_runs}{Integer value of the number of alterations to be done for
method = random.}

\item{layers_different}{Should layers from layers_map be changed individually?
If \code{TRUE} this allows for different changes at different depths.}
}
\description{
An S3 class for run maps generated by a call to run_map() to
be used in alternate().
}
