% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_soilphys.R
\name{cfp_soilphys}
\alias{cfp_soilphys}
\title{cfp_soilphys}
\usage{
cfp_soilphys(soilphys, id_cols)
}
\arguments{
\item{soilphys}{A data.frame with (at least) the following columns:
\describe{
\item{upper (cm) }{The upper bound of each step.}
\item{lower (cm)}{The lower bound of each step.}
\item{gas}{The gas of that step.}
\item{DS (\eqn{m^2 s^-1})}{The specific diffusion coefficient of that gas in that step.}
\item{c_air (\eqn{mol m^-3})}{The number density of air in that step.}
\item{any of \code{id_cols}}{All id_cols that identify one profile uniquely.}

}}

\item{id_cols}{Column names in data.frame that uniquely identify each profile.}
}
\value{
cfp_soilphys
}
\description{
Create and validate cfp_soilphys objects. A cfp_soilphys object has the following
characteristics. Each profile is uniquely identifiable by its \code{id_cols} without duplicates.
The data is upper/lower consistent, meaning that the steps of each profile cover a range
from min(lower) to max(upper) without gaps, overlap or duplicates.
}
