% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_soilphys.R
\name{check_soilphys}
\alias{check_soilphys}
\title{check_soilphys}
\usage{
check_soilphys(df, extra_vars = c())
}
\arguments{
\item{df}{(dataframe) the soilphys dataframe}

\item{vars}{(character vector) column names of additional variables to be checked.}
}
\value{

}
\description{
This function analyses the soilphys dataframe before the flux calculation. It presents a
warning, if there are variables missing and also looks for suspicious patterns that suggest an error in the interpolation made by discretize_depth.
}
\examples{

}
