% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_models.R
\name{combine_models}
\alias{combine_models}
\alias{combine_models.cfp_altres}
\alias{combine_models.list}
\title{Combine Models}
\usage{
combine_models(x)

\method{combine_models}{cfp_altres}(x)

\method{combine_models}{list}(x)
}
\arguments{
\item{x}{A list of models, must inherit from \code{\link[=cfp_dat]{cfp_dat()}}}
}
\description{
Combinea list of multiple models or \code{\link[=cfp_dat]{cfp_dat()}} objects into a single object.
}
\examples{
mod1 <- filter(base_dat, site == "site_a")
mod2 <- filter(base_dat, site == "site_b")
combine_models(list(mod1, mod2))
}
