% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D0_massman.R
\name{D0_massman}
\alias{D0_massman}
\title{D0_Massman}
\usage{
D0_massman(gas, Temp, p)
}
\arguments{
\item{gas}{(character) One of "CO2","CH4","N2O","O2","N2"}

\item{p}{(numeric) pressure in hpa}

\item{t}{(numeric) temperature in °C}
}
\value{
D0 in µmol/m ??
}
\description{
This function calculates the Diffusion coefficients of different gases in air at set
temperature and pressure. The functions are taken from Massman 1998 "A REVIEW OF THE MOLECULAR DIFFUSIVITIES
H2O, CO2, CH4, CO, O3, SO2, NH4, N2O, NO, AND NO2 IN AIR, O2 AND N2 NEAR STP".
}
