% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depth_structure.R
\name{depth_structure}
\alias{depth_structure}
\alias{depth_structure.cfp_layered_profile}
\alias{depth_structure.cfp_profile}
\alias{depth_structure.cfp_dat}
\title{Unique layers depths}
\usage{
depth_structure(x, id_cols = NULL, ...)

\method{depth_structure}{cfp_layered_profile}(x, id_cols = NULL, ...)

\method{depth_structure}{cfp_profile}(x, id_cols = NULL, ...)

\method{depth_structure}{cfp_dat}(x, id_cols = NULL, structure_from = NULL, ...)
}
\arguments{
\item{x}{An object to get general structure of.}
}
\description{
Get the unique layers or depths, i.e. the backbone
of an object given a set of identifying columns.
}
