% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efflux.R
\name{efflux}
\alias{efflux}
\title{efflux}
\usage{
efflux(x, ...)
}
\arguments{
\item{x}{A cfp_pfres or cfp_fgres model result.}

\item{...}{Further arguments passed on for cfp_fgres efflux calculation.}
}
\description{
Calculate or extract the soil/atmosphere efflux
from cfp_pfres or cfp_fgres model results.
}
