% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfp_altres.R, R/cfp_dat.R, R/cfp_fgmod.R,
%   R/cfp_pfmod.R, R/cfp_run_map.R, R/extractors.R
\name{cfp_og_model}
\alias{cfp_og_model}
\alias{cfp_id_cols}
\alias{get_soilphys}
\alias{get_gasdata}
\alias{get_layers_map}
\alias{cfp_gases}
\alias{cfp_modes}
\alias{cfp_param}
\alias{cfp_funs}
\alias{cfp_zero_flux}
\alias{cfp_zero_limits}
\alias{cfp_DSD0_optim}
\alias{cfp_evenness_factor}
\alias{cfp_known_flux_factor}
\alias{cfp_runmap_type}
\alias{cfp_params_df}
\alias{cfp_n_runs}
\alias{extractors}
\title{extractors}
\usage{
cfp_og_model(x)

cfp_id_cols(x)

get_soilphys(x)

get_gasdata(x)

get_layers_map(x)

cfp_gases(x)

cfp_modes(x)

cfp_param(x)

cfp_funs(x)

cfp_zero_flux(x)

cfp_zero_limits(x)

cfp_DSD0_optim(x)

cfp_evenness_factor(x)

cfp_known_flux_factor(x)

cfp_runmap_type(x)

cfp_params_df(x)

cfp_n_runs(x)
}
\arguments{
\item{x}{An object from which to extract the information.}
}
\description{
These functions help extract information from
different objects that can be defined in ConFluxPro.
}
\section{Functions}{
\itemize{
\item \code{cfp_og_model()}: og_model

\item \code{cfp_id_cols()}: id_cols

\item \code{get_soilphys()}: get_soilphys

\item \code{get_gasdata()}: get_gasdata

\item \code{get_layers_map()}: get_layers_map

\item \code{cfp_gases()}: gases

\item \code{cfp_modes()}: modes

\item \code{cfp_param()}: param

\item \code{cfp_funs()}: funs

\item \code{cfp_zero_flux()}: zero_flux

\item \code{cfp_zero_limits()}: zero_limits

\item \code{cfp_DSD0_optim()}: DSD0_optim

\item \code{cfp_evenness_factor()}: evenness_factor

\item \code{cfp_known_flux_factor()}: known_flux_factor

\item \code{cfp_runmap_type()}: runmap_type

\item \code{cfp_params_df()}: params_df

\item \code{cfp_n_runs()}: n_runs

}}
