% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_flux.R
\name{fg_flux}
\alias{fg_flux}
\alias{calculate_flux}
\title{fg_flux}
\usage{
fg_flux(x, ...)

calculate_flux(
  gasdata,
  soilphys,
  layers_map,
  gases,
  modes,
  param,
  funs,
  id_cols = NULL
)
}
\arguments{
\item{x}{A \code{cfp_dat} object with all the necessary input datasets.}

\item{...}{Additional arguments passed to fg_flux.cfp_fgmod. Can be of the following:}

\item{gasdata}{A cfp_gasdata object created by running \code{cfp_gasdata()}.}

\item{soilphys}{A cfp_soilphys object created by running \code{cfp_soilphys()}.}

\item{layers_map}{A cfp_layers_map object created by running \code{cfp_layers_map}.}

\item{gases}{(character) A character vector defining the gases for which
gluxes shall be calulated.}

\item{modes}{(character) A character vector specifying mode(s) for dcdz
calculation. Can be "LL","LS","EF".}

\item{param}{(character) A vector containing the the parameters of soilphys,
for which means should be calculated, must contain "c_air" and "DS", more
parameters help interpretation}

\item{funs}{(character) A vector defining the type of mean to be used. One of
"arith" or "harm"}

\item{id_cols}{(character vector) The names of the columns that together
uniquely identify one profile.}
}
\description{
This function takes a valid input dataset in cfp_dat
and calculates fluxes accordingly. Fluxes are calculated for each layer defined
in layers map and are given in mumol/m^2/s.
}
