% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_flux.R
\name{fg_flux}
\alias{fg_flux}
\title{fg_flux}
\usage{
fg_flux(x, ...)
}
\arguments{
\item{x}{A valid \code{cfp_dat} object. This contains all the necessary input
data.}

\item{gases}{(character) A character vector defining the gases for which
gluxes shall be calulated.}

\item{modes}{(character) A character vector specifying mode(s) for dcdz
calculation. Can be "LL","LS","EF".}

\item{param}{(character) A vector containing the the parameters of soilphys,
for which means should be calculated, must contain "c_air" and "DS", more
parameters help interpretation}

\item{funs}{(character) A vector defining the type of mean to be used. One of
"arith" or "harm"
@export}
}
\description{
This function takes a valid input dataset in cfp_dat
and calculates fluxes accordingly. Fluxes are calculated for each layer defined
in layers map and are given in mumol/m^2/s.
}
