% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fg_flux.R
\name{fg_flux}
\alias{fg_flux}
\alias{fg_flux.cfp_dat}
\alias{fg_flux.cfp_fgres}
\alias{fg_flux.cfp_fgmod}
\title{Flux gradient method}
\usage{
fg_flux(x, ...)

\method{fg_flux}{cfp_dat}(
  x,
  ...,
  gases = NULL,
  modes = "LL",
  param = c("c_air", "DS"),
  funs = c("arith", "harm")
)

\method{fg_flux}{cfp_fgres}(x, ...)

\method{fg_flux}{cfp_fgmod}(x, ...)
}
\arguments{
\item{x}{A \code{cfp_dat} object with all the necessary input datasets.}

\item{...}{
  Arguments passed on to \code{\link[=cfp_fgmod]{cfp_fgmod}}
  \describe{
    \item{\code{gases}}{(character) A character vector defining the gases for which
fluxes shall be calulated.}
    \item{\code{modes}}{(character) A character vector specifying mode(s) for dcdz
calculation. Can be \code{"LL"},\code{"LS"},\code{"EF"}.
\describe{
\item{LL}{local linear approach: within each layer a linear model is
evaluated of concentration over the depth.}
\item{LS}{linear spline approach: A linear spline is fitted over the
complete profile with nodes at the layer intersections.}
\item{EF}{exponential fit approach: An exponential function of form
y=a+b*x^c is fit of concentration against depth. Using the first derivative of
that function the concentration gradient is evaluated for each layer.}
}}
    \item{\code{param}}{(character) A vector containing the the parameters of soilphys,
for which means should be calculated, must contain "c_air" and "DS", more
parameters may help interpretation.}
    \item{\code{funs}}{(character) A vector defining the type of mean to be used for
each parameter in \code{param}. One of "arith" or "harm".}
  }}
}
\description{
This function takes a valid input dataset in cfp_dat
and calculates fluxes accordingly. Fluxes are calculated for each layer defined
in layers map and are given in mumol/m^2/s.
}
