% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux}
\alias{flux}
\title{Re-run model}
\usage{
flux(x)
}
\arguments{
\item{x}{A valid \code{cfp_fgmod} or \code{cfp_pfmod} object.}
}
\description{
A function to either run \code{fg_flux()} or \code{pro_flux()}
models from valid \code{cfp_fgmod} or \code{cfp_pfmod} objects.
}
