% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harm.R
\name{harm}
\alias{harm}
\title{harm}
\usage{
harm(x, w = 1, na.rm = F)
}
\arguments{
\item{x}{(numeric vector)}

\item{w}{(numeric vector) optional vector of weights corresponding to x.
Default is 1 for all.}

\item{na.rm}{(logical) If TRUE, then NA values are omitted and the mean
calculated with the remaining values. If FALSE (default) then returns NA if
x contains NA values.}
}
\value{
(numeric) harmonic mean of x
}
\description{
This function calculates harmonic mean of a vector and can be
 used analogous to the base functions mean() or median()
}
\examples{
{
 harm(c(1:10))
 harm(c(1:10),c(10:1))
}


}
