% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lin_extrap.R
\name{lin_extrap}
\alias{lin_extrap}
\title{lin_extrap}
\usage{
lin_extrap(x, y, x_new)
}
\arguments{
\item{x}{(numeric) Two values of x}

\item{y}{(numeric) Two values of y (same order as x)}

\item{x_new}{(numeric) x-values to be extrapolated to.}
}
\value{
y_new
}
\description{
This function linearly extrapolates based on
two value-pairs (x1|y1),(x2|y2) to a third point of x (x_new).
}
\examples{
{
lin_extrap(c(3,4),c(10,20),5)
lin_extrap(c(3,4),c(10,20),c(5,-5))
}
}
