% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offset_override.R
\name{offset_override}
\alias{offset_override}
\title{offset_override}
\usage{
offset_override(corr_map, SAMPLE_NO, section)
}
\arguments{
\item{corr_map}{(dataframe) The corr_map dataframe created by offset_subsetting}

\item{SAMPLE_NO}{(character) A vevctor of the SAMPLE_NO to change section}

\item{section}{(vector, numeric) A vector in the same order as SAMPLE_No specifying the section to add to.}
}
\value{
corr_map (dataframe) The corr_map dataframe used in the offset_correction function with overriden SAMPLE_NO
}
\description{
This function helps overriding the Date-based subsetting of offset_subsetting.
It modifies an existing corr_map by editing the section of a list of SAMPLE_NO. Check the section numbering of corr_map first.
If a new section is manually introduced this way, a mode will have to be added manually.
}
\examples{
offset_override(corr_map = corr_map,
SAMPLE_NO = c(3023,5612,405,"blackforest27"),
section = c(6,2,9,3))

}
