% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_extract.R
\name{pf_efflux}
\alias{pf_efflux}
\title{proflux_extract}
\usage{
pf_efflux(PROFLUX, ...)
}
\arguments{
\item{PROFLUX}{The PFres object returned by pro_flux()}

\item{id_cols}{The id_cols used in the function call(only needed
if PROFLUX is not a PFres object).}
}
\value{
data.frame
}
\description{
Utility function for the fast extraction and aggregation of
parameters exported by pro_flux
}
