% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro_flux.R
\name{pro_flux}
\alias{pro_flux}
\title{pro_flux}
\usage{
pro_flux(
  gasdata,
  soilphys,
  prod_depth,
  storage_term = F,
  zero_flux = T,
  highlim,
  lowlim,
  id_cols
)
}
\arguments{
\item{gasdata}{(dataframe)}

\item{soilphys}{(dataframe)}

\item{storage_term}{(logical) Should changes in storage be accounted for? Default is F.
Only works if data is present in a temporal dimension as well and is probably only representative for a high temporal resolution (hours).}

\item{zero_flux}{(logical) Applies the zero-flux boundary condition (T)? If FALSE, the first value in X
represents the incoming flux to the lowest layer.}

\item{target_depth}{(numeric vector) This vector determines the depths of the interfaces between
different production values to be fitted. This allows for an adaptive model with an individual number variables. So for three different production values, two depths that mark the intersection must be given.}
}
\description{
This implements an inverse modeling approach which optimizes vertically resolved production (or consumption)
of the gases in question to fit a modeled concentration profile to observed data.
}
\examples{
pro_flux(gasdata,
soilphys,
target_depth,
storage_term = F,
zero_flux = T)

}
