% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_optim.R
\name{prod_optim}
\alias{prod_optim}
\title{prod_optim}
\usage{
prod_optim(X, height, DS, C0, pmap, cmap, conc, dstor, zero_flux = T)
}
\arguments{
\item{X}{(numeric vector) specifying the productions to be optimised}

\item{C0}{(numeric) The concentration at the bottom of the lowermost step}

\item{pmap}{(integer vector) assigning a production from X to each step}

\item{cmap}{(integer vector) assigning the modelled concentrations to the observed concentrations
as there can be multiple observations per depth}

\item{conc}{(numeric) the observed concentrations}

\item{dstor}{(numeric) storage changes per step}

\item{zero_flux}{(logical) Applies the zero-flux boundary condition(T)? If FALSE, the first value in X
represents the incoming flux to the lowest layer.}
}
\description{
This is the optimiser-function that is minimised for the inverse,
production based model. It takes as input a vector of the influx, as well as the
values of the production to be optimised.
}
\examples{

}
