% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_alternate.R
\name{proflux_alternate}
\alias{proflux_alternate}
\title{proflux_alternate}
\usage{
proflux_alternate(
  PROFLUX,
  params,
  facs,
  params_map,
  topheight_var = FALSE,
  sensitivity_analysis = FALSE,
  n_runs = 0,
  no_confirm = FALSE,
  DSD0_formula,
  return_raw = FALSE,
  write_db = FALSE,
  con = NULL,
  error_funs,
  error_args
)
}
\arguments{
\item{PROFLUX}{An object of class PFres, the result of a call to
\code{\link{pro_flux}}}

\item{params}{A vector of characters which parameters to alternate. Can be
  the character of any column in \code{soilphys} or "topheight". If
  "topheight" is called, the height of the soil/atmosphere interface will be
  altered. This means, that the corresponding values in \code{gasdata},
  \code{soilphys} and \code{layers_map} will be adjusted.

  The parameter(s) provided will be varied along the \code{facs} provided.
  Then, \code{\link{complete_soilphys}} is run. This means, that it does not
  make sense to alter, e.g. "AFPS" as it is re-calculated in that function
  call.}

\item{facs}{A vector of factors each of params will be multiplied by. This
may be something like \code{seq(0.9,1.1,0.05)} if values should be varied
between 90 \% and 110 \% of their original value.}

\item{params_map}{A \code{data.frame} containing any of \code{id_cols} of
PROFLUX, as well as \code{upper} and \code{lower}. For each layer enclosed
by "upper" and "lower", the parameters in "params" will be varied
independently. This means, that the more layer per combination of id_cols
are provided, the more variations will be tested out.}

\item{topheight_var}{Either FALSE (default) or a vector of numerics that define
how much the topheight should be varied in cm. Usually this corresponds to
the humus layer of a soil. The values are added to the original topheight.
Negative values are subtracted. If this results in a layer of height 0
(or negative height), it is removed from the model and a warning message displayed.}

\item{sensitivity_analysis}{How should the different values of \code{params}
and \code{facs} be combined? Can be one of:
\describe{
\item{FALSE}{(default). Here, all possible combinations of \code{params},
             \code{facs} and the respective layers of \code{params_map}
             are run.}
\item{OAT}{One At a Time. This means, that only one of \code{params} is
           changed at a time for all layers, all the others will have
           a \code{facs} of 1.}
\item{OATL}{One At a Time Layer. Here, \code{OAT} is expanded to change
            only one parameter \emph{and} one layer at a time.}
\item{RAND}{Within the range of facs, values of params are randomly
varied. This is repeated \code{n} times}
\item{RANDL}{The same is \code{RAND} but also allows for different values
between layers.}
}}

\item{n_runs}{The number of runs to do for
\code{sensitivity_analysis == c("RAND","RANDL")}.}

\item{no_confirm}{If the function is used in a script, set this to TRUE to
confirm the number of profiles without separate user input. Please ensure
separately if the function will evaluate in a reasonable amount of time.}

\item{DSD0_formula}{(character) A character vector defining the way DSD0
should be calculated. Must refer to existing columns in soilphys. See
examples below.}

\item{return_raw}{If TRUE, the complete model results are returned as a list
instead of the error summary.}

\item{write_db}{If \code{TRUE}, the results (either the error parameters from error_funs
or the raw PFres results) are written in table in the database defined in con.
Runs are run together at 25 runs each to improve performance while preventing memory crashes.}

\item{con}{The database connection to be used when write_db is \code{TRUE}.}

\item{error_funs}{A named list of functions that take PROFLUX (+ further arguments)
as input and return a \code{data.frame} of at least one column: NRMSE, as
well as further columns that may represent different parameters calculated
in parallel (e.g. NRMSE for different gases or id_cols).}

\item{error_args}{A list of one named list of function arguments per function
provided in error_funs. Note, that the arguments must be named. The names of the
wrapping list must follow the names and order of \code{error_funs}. Additional
data needed in these functions (e.g. reference \code{EFFLUX} values) must be
provided here.}
}
\description{
Functions to extract datasets from the result of a
  \code{\link{pro_flux}} model, a PFres object, to modify parameters and
  rerun the model.
}
