% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proflux_summarise.R
\name{proflux_summarise}
\alias{proflux_summarise}
\alias{proflux_summarize}
\alias{proflux_summarise.data.frame}
\alias{proflux_summarise.PFres}
\title{proflux_summarise}
\usage{
proflux_summarise(PROFLUX, ...)

proflux_summarize(PROFLUX, ...)

\method{proflux_summarise}{data.frame}(PROFLUX, layers_map, id_cols, ...)

\method{proflux_summarise}{PFres}(PROFLUX, ...)
}
\arguments{
\item{PROFLUX}{Either an object of class \code{\link{PFres}}
(see \code{\link{pro_flux}}), or a data.frame following the same convention.}
}
\description{
Summarises key parameters of a PFres object
(produced by a call to \code{\link{pro_flux}}) based on the id_cols of its layers_map.
}
