% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_map.R
\name{run_map}
\alias{run_map}
\title{run_map}
\usage{
run_map(
  x,
  params = list(),
  method = NULL,
  type = NULL,
  n_runs = NULL,
  layers_different = FALSE,
  topheight_adjust = FALSE
)
}
\arguments{
\item{params}{A named list of numeric vectors. Names indicate column names
in soilphys, vectors either distinct values (method permutation) or limits (method random).}

\item{type}{A vector of length param indicating what the values in params
represent. One of
\describe{
\item{abs}{Absolute values that are applied as-is.}
\item{factor}{Factors to be multiplied with the original values.}
\item{addition}{Factors to be added to the original values.}
}}

\item{n_runs}{Integer value of the number of alterations to be done for
method = random.}

\item{layers_different}{Should layers from layers_map be changed individually?
If \code{TRUE} this allows for different changes at different depths.}

\item{topheight_adjust}{(logical) If the proposed change in topheight is larger
than the highest layer in soilphys, should the limits be automatically
adjusted per id_cols indivdually? Default is FALSE, which leads to an error in that
case.}
}
\description{
Create a run_map for model alteration in \code{alternate()}.
}
