% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_models.R
\name{scale_min_median}
\alias{scale_min_median}
\title{Scale a vector by its min and median}
\usage{
scale_min_median(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
\code{x} scaled between min and median of \code{x}.
}
\description{
Scale a vector between its minimum and median.
}
\examples{
scale_min_median(1:10)

}
