% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/season.R
\name{season}
\alias{season}
\title{ConFluxPro}
\usage{
season(d)
}
\arguments{
\item{d}{(Date)}
}
\value{
m (character)
}
\description{
A simple function to return a character (-vector) of the season from a Date (-vector).
Months:
spring 3-5;
summer 6-8;
fall 9-11;
winter 12-2;
}
\examples{
d <- as.Date("2015-10-15")
s <- season(d)

d <- as.Date(c("1955-01-15","1985-06-15","2015-10-15"))
s <- season(d)

}
