% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series_correction.R
\name{series_cleaner}
\alias{series_cleaner}
\title{series_correction}
\usage{
series_cleaner(df, mode = "data")
}
\arguments{
\item{df}{(dataframe)}

\item{mode}{(character) "data" returns the corrected dataset. "plots" returns
before / after plots.}
}
\value{
dataframe
}
\description{
This function corrects time series of gas concentrations within
  a certain depth at one site. The basic assumption is that gas data of
  different samplers within a depth are strictly ordered due to spatial
  heterogeneity or differences in the exact installation depth, i.e. always
  one sampler shows the highest concentration, so that the difference of each
  sampler on each day to the respective mean of all samplers is more or less
  the same.

  This function does not apply if the data either has only one observation
  per depth or if the data does not show any intrinsic ordering inside each
  depth. Check these preconditions first.

  The function outputs a dataframe that has the same structure as the input,
  so that the original input can be removed. The function can be used using
  dplyr group_modify()-function
}
\examples{
{
data("gasdata")

gasdata \%>\%
  filter(depth == 0) \%>\% #only to make example smaller
  mutate(MST_ID = repetition,
         depth_cat = depth) \%>\%
  group_by(site,depth_cat) \%>\%
  group_modify(~series_cleaner(.x))}

}
\seealso{
Other gasdata: 
\code{\link{balance_correction}()},
\code{\link{offset_correction}()},
\code{\link{offset_override}()},
\code{\link{offset_subsetting}()}
}
\concept{gasdata}
