% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_calc_indices.R
\name{sobol_calc_indices}
\alias{sobol_calc_indices}
\title{sobol_calc_indices}
\usage{
sobol_calc_indices(Y, effect_cols, id_cols = character(), run_map)
}
\arguments{
\item{Y}{A data.frame with the desired effect parameter(s) of the model output. E.g.: efflux.}

\item{effect_cols}{character vector of the column names in Y giving the target effects for which sobol indices
should be calculated}

\item{id_cols}{character vector of column names specifying grouping columns. If the original model had
id_cols (e.g. different sites or gases) you may want to calculate indices per id_cols combination separately.}

\item{run_map}{A data.frame created by \code{run_map()} with the necessary information
how the data is to be changed with each distinct \code{run_id}.}
}
\description{
From any result parameter and its corresponding run_map calculate
first-order and total sobol indices using the Azzini (2021) method.
}
