% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol_run_map.R
\name{sobol_run_map}
\alias{sobol_run_map}
\title{sobol_run_map}
\usage{
sobol_run_map(x, ...)
}
\arguments{
\item{x}{An object of class cfp_run_map created by a call to run_map().}

\item{...}{Additional arguments passed on to \link{run_map}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Modify an existing run_map for sobol indice estimation.
}
