% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilphys_layered.R
\name{soilphys_layered}
\alias{soilphys_layered}
\title{soilphys_layered}
\usage{
soilphys_layered(
  soilphys,
  layers_map,
  param = c("TPS", "SWC", "AFPS", "T", "p", "DSD0", "D0", "DS"),
  funs = c("arith", "arith", "arith", "arith", "arith", "harm", "harm", "harm")
)
}
\arguments{
\item{soilphys}{(dataframe) the soilphys dataframe}

\item{layers_map}{(dataframe) containing the following parameters: "Plot",layer"=name of the layer;
"upper"=upper limit of layer in cm; "lower" = lower limit of the layer in cm; Must be the same
dataframe as the input of the same name in dcdz_layered().}

\item{param}{(character vector) A vector containing the names of the variables in soilphys to be carried over}

\item{funs}{(character vector) A vector defining the type of mean to be used. One of "arith" or "harm"}
}
\value{

}
\description{
This function calculates (harmonic-) means of the soilphys dataframe per layer
for the flux calculation.
}
\examples{

}
